refactored/syncboard_backend/backend/requirements.txt
fastapi
uvicorn
pydantic
numpy
scikit-learn
requests
python-multipart
python-dotenv

# AI API clients
openai

# Content ingestion dependencies
yt-dlp
pypdf
beautifulsoup4
python-docx

# Image processing
Pillow
pytesseract
